Rename main literal types (`Lit` -> `Literal`, `Foo` -> `FooLit`)
Without the `Lit` suffix, it's a bit confusing for people importing
these types. Especially for the upcoming string literals. I didn't
choose `Literal` as suffix as it might be a bit long then. And "lit"
is already a common abbreviation in Rust world.

The main type was renamed from `Lit` to `Literal` though. I feel using
the long form here is better.