Add several derives for `Eq` and `Copy`
So regarding `[Partial]Eq`: it makes sense that a user of this library
wants to compare two literals. There is a well defined equality
relation here (i.e. basically the input string). All types currently
implement `PartialEq` in a way the is consistent with that.

Though it's sometimes technically inefficient as all fields are
compared even when a subset would suffice. But oh well, that's fine
for now.