Move lots of parsing code into non-generic functions
This should improve compile times as the main logic can be compiled
once (when you see "Compiling litrs") and not as part of your crate's
instantiation. The latter should be dealt with by incremental
compilation, but it's not always perfect and generally this change
should make it easier for the Rust compiler.

It might also reduce binary sizes (which is hardly relevant for proc
macros).

That said, I have not tested any of that as it's not trivial. I know,
not great practice :/