Remove CR LF normalization to align with spec
The specification now says that CR LF normalization is part of the
pre-processing prior to tokenization. Since it's not part of the lex
grammar, this commit removes it from the parsing code as well. CR are
simply disallowed fully now.

This is technically a breaking change, but unlikely to be noticed by
any real world input. In proc macros, all input has been normalized by
the Rust compiler anyway, so only very weird input (CR CR LF) would
have been accepted previously, but not after this commit.

See:
- https://github.com/rust-lang/reference/pull/1944
- https://github.com/rust-lang/reference/commit/fa56fdba0e9dba35eb29d11c95c7a009ed67cb35
- https://github.com/rust-lang/reference/commit/27e1ec97a75267d3c9efb9c91c7509eff98d11db