Refactor `unescape` to not take `offset` parameter anymore
This was a source of bugs and was not very nice anyway. It was just
used for error spans, which was confusing. This span offsetting is
better done by the calling function, which this commit does. This commit
fixes one bug as well: for `InvalidStartOfUnicodeEscape` the wrong span
was returned unless the escape was the first thing in the string.